scripts/requirements.txt
open_clip_torch>=2.24.0
torch>=2.0.0
Pillow>=10.0.0
numpy>=1.24.0
